Cherry-pick inanna doc additions; move entity Views to data/<feature>/
Aligns hathor docs with three doc-level commits inanna landed yesterday
(35fa339, b5f2a99, 236420c) and adopts inanna's `data/<feature>/` View
convention by moving the four entity-tied page files out of `pages/`.

- Move VehicleTypeView, VehicleTypeDetails, DeckPlanView,
  DeckPlanDetailsView from `src/pages/` into their `src/data/<feature>/`
  folders. Generic infra (GenericDataViewPage, GenericDataEditPage,
  GenericDetailsPage) and Home stay in `pages/`.
- DEV_GUIDE.md: add Folder semantics section; fix stale theme/icon
  paths (theme/, utils/iconLoaderUtils.ts); replace inanna trailer with
  hathor's real examples; drop the inanna-copy comment.
- OPEN_QUESTIONS.md: new at repo root, hathor-real design ambiguities
  only (5 items, neutral framing).
- CLAUDE.md: split debt — design ambiguities moved to OPEN_QUESTIONS.md;
  pure dead-code chores stay; fix path references after View migration.
- README.md: pointer line mentions project layout + OPEN_QUESTIONS.
- .prettierignore: add `.vite` (vite dep cache) to unblock check.

Closes #53
