+++ src/data/deck-plans/DeckPlanView.tsx
@@ -1,8 +1,8 @@
-import { getDeckPlanViewConfig } from '../data/deck-plans/deckPlanViewConfig.ts';
-import GenericDataViewPage from './GenericDataViewPage.tsx';
-import type { DeckPlan } from '../data/vehicle-types/vehicleTypeTypes.ts';
-import { useUrlFilters } from '../hooks/useUrlFilters.ts';
-import type { OrderBy } from '../data/deck-plans/useDeckPlans.ts';
+import { getDeckPlanViewConfig } from './deckPlanViewConfig.ts';
+import GenericDataViewPage from '../../pages/GenericDataViewPage.tsx';
+import type { DeckPlan } from '../vehicle-types/vehicleTypeTypes.ts';
+import { useUrlFilters } from '../../hooks/useUrlFilters.ts';
+import type { OrderBy } from './useDeckPlans.ts';
 import { useNavigate } from 'react-router-dom';
 
 export default function DeckPlanView() {
